moving common validations from cmd to pkg/workflows
